create the output analysing functions

+++ anduino_deployments/kws/Kws/test_input.h
@@ -1,1 +1,2 @@
-extern const float input_val[];
+#include "stdint.h"
+extern const int8_t input_val[];
